# array join: avoid strcat, speedup from O(N^2) to O(N)