@avih 2022-02-16
For reference, my test program (a bit more than 80 chars per element...):
```js
var s = "x";
while (s.length < 80)
    s = s + s;

var a = []
for (var i = 0; i < 2000; i++)
    a.push("" + i + " " + s);

var t0 = Date.now()
a = a.join("\n");
print("ms: " + (Date.now() - t0));
```